CoDS-COMAD cleanup and consolidation
This commit tidies up deadline mode spaghetti written aroudn CoDS-COMAD
submission around July 20. Some of the repeat codes, non DRY aspects and
break of convention are corrected.

There was a possibility of an infinite loop in pib.cli.translate_pib,
which created issues once the Punkt tokenizer was introduced. The
routine is not embedded with additional assertions to ensure that at
least one step of an entry is taken in the loop and that no empty batch
is returned, to ensure no silent failure in other parts.

This could be perhaps what led to issues (wrong offset) and wrong
translations in PIB round 2 during CoDs COMAD, eventually fixed by
adding article information in the pipeline. A stronger queue based
translate mechanism with a buffer in place could be implemented for
better batching and faster translation of PIB articles.

- lazy_loads -> lazy_load bugfix
- Monolingual corpus export made better
- Adding shashank's corpus comparison script
- Possibly leftoever from CoDS COMAD
- bumping model to M2EN3
- minor; converting listcomp to imperative
- Adjusting script, adding model_dir
- Bugfixes in translation routine
- Reflecting entry update with every index increment
- Adding language restriction capability to store-retrieved